requirements.txt
azure_storage==0.36.0
numpy==1.23.5
pandas==1.5.2
PyPDF2==2.11.2
spacy==3.4.3
spacy_streamlit==1.0.4
streamlit==1.13.0
tabulate==0.9.0

